feat(bivariate-matrix): add ability for bivariate matrix to load dymically generated tiles
